Project updated

+++ src/tests/tests_unit_create_datasets.py
@@ -140,7 +140,7 @@ def logit(h):
     ''' Logistic from activation h '''
     p = 1.0 / (1.0 + np.exp(-h))
     r = np.random.rand(len(p))
-    y = (r < p).astype('int')
+    y = (r < p).astype('float')
     return y
